trackdissect/requirements.txt
fastapi
uvicorn
python-multipart
librosa
numpy
soundfile
anthropic
google-genai
demucs
